run spinner when saving targets
